+++ callback.php
@@ -31,17 +31,19 @@ const TrackerStatus_Closed = 4;
 const TrackerStatus_ForceSave = 6;
 const TrackerStatus_CorruptedForceSave = 7;
 
+$plugin = OnlyofficePlugin::create();
+
 if (isset($_GET["hash"]) && !empty($_GET["hash"])) {
-    $response_array;
+    $callbackResponseArray = [];
     @header( 'Content-Type: application/json; charset==utf-8');
     @header( 'X-Robots-Tag: noindex' );
     @header( 'X-Content-Type-Options: nosniff' );
 
     list ($hashData, $error) = Crypt::ReadHash($_GET["hash"]);
     if ($hashData === null) {
-        $response_array["status"] = "error";
-        $response_array["error"] = $error;
-        die(json_encode($response_array));
+        $callbackResponseArray["status"] = "error";
+        $callbackResponseArray["error"] = $error;
+        die(json_encode($callbackResponseArray));
     }
 
     $type = $hashData->type;
@@ -76,24 +78,26 @@ if (isset($_GET["hash"]) && !empty($_GET["hash"])) {
 
     switch($type) {
         case "track":
-            $response_array = track();
-            die (json_encode($response_array));
+            $callbackResponseArray = track();
+            die (json_encode($callbackResponseArray));
         case "download":
-            $response_array = download();
-            die (json_encode($response_array));
+            $callbackResponseArray = download();
+            die (json_encode($callbackResponseArray));
         default:
-            $response_array["status"] = "error";
-            $response_array["error"] = "404 Method not found";
-            die(json_encode($response_array));
+            $callbackResponseArray["status"] = "error";
+            $callbackResponseArray["error"] = "404 Method not found";
+            die(json_encode($callbackResponseArray));
     }
 }
 
 /**
  * Handle request from the document server with the document status information
  */
-function track() {
+function track(): array
+{
     $result = [];
 
+    global $plugin;
     global $courseCode;
     global $userId;
     global $docId;
@@ -111,6 +115,32 @@ function track() {
     if ($data === null) {
         $result["error"] = "Bad Response";
         return $result;
+    }
+
+    if (!empty($plugin->get("jwt_secret"))) {
+
+        if (!empty($data["token"])) {
+            try {
+                $payload = \Firebase\JWT\JWT::decode($data["token"], $plugin->get("jwt_secret"), array("HS256"));
+            } catch (\UnexpectedValueException $e) {
+                $result["status"] = "error";
+                $result["error"] = "403 Access denied";
+                return $result;
+            }
+        } else {
+            $token = substr(getallheaders()[AppConfig::JwtHeader()], strlen("Bearer "));
+            try {
+                $decodeToken = \Firebase\JWT\JWT::decode($token, $plugin->get("jwt_secret"), array("HS256"));
+                $payload = $decodeToken->payload;
+            } catch (\UnexpectedValueException $e) {
+                $result["status"] = "error";
+                $result["error"] = "403 Access denied";
+                return $result;
+            }
+        }
+
+        $data["url"] = isset($payload->url) ? $payload->url : null;
+        $data["status"] = $payload->status;
     }
 
     $status = $data["status"];
@@ -185,13 +215,27 @@ function track() {
 /**
  * Downloading file by the document service
  */
-function download() {
+function download()
+{
+    global $plugin;
     global $courseCode;
     global $userId;
     global $docId;
     global $groupId;
     global $sessionId;
     global $courseInfo;
+
+    if (!empty($plugin->get("jwt_secret"))) {
+        $token = substr(getallheaders()[AppConfig::JwtHeader()], strlen("Bearer "));
+        try {
+            $payload = \Firebase\JWT\JWT::decode($token, $plugin->get("jwt_secret"), array("HS256"));
+
+        } catch (\UnexpectedValueException $e) {
+            $result["status"] = "error";
+            $result["error"] = "403 Access denied";
+            return $result;
+        }
+    }
 
     if (!empty($docId) && !empty($courseCode)) {
         $docInfo = DocumentManager::get_document_data_by_id($docId, $courseCode, false, $sessionId);
@@ -215,16 +259,9 @@ function download() {
 
 /**
  * Method checks access rights to document and returns permissions
- * 
- * @param array $docInfo - identifier of document
- * @param int $userId - identifier of user
- * @param string $courseCode - identifier of course
- * @param int $groupId - identifier of group or null if file out of group
- * @param int $sessionId - identifier of session
- * 
- * @return array
  */
-function getPermissions($docInfo, $userId, $courseCode, $groupId, $sessionId) {
+function getPermissions(array $docInfo, int $userId, string $courseCode, int $groupId = null, int $sessionId = null): array
+{
     $isAllowToEdit = api_is_allowed_to_edit(true, true);
     $isMyDir = DocumentManager::is_my_shared_folder($userId, $docInfo["absolute_parent_path"], $sessionId);