+++ lang/english.php
@@ -1,12 +1,13 @@
 <?php
 /**
  * @author ASENSIO SYSTEM SIA
- * 
+ *
  */
 $strings["plugin_title"] = "ONLYOFFICE";
 $strings['plugin_comment'] = "ONLYOFFICE connector allows you to view, edit, and collaborate on text documents, spreadsheets, and presentations within Chamilo using ONLYOFFICE Docs.";
-$strings["enableOnlyofficePlugin"] = "Enable";
-$strings["documentServerUrl"] = "Document Editing Service address";
+$strings["enable_onlyoffice_plugin"] = "Enable";
+$strings["document_server_url"] = "ONLYOFFICE Docs address";
+$strings["jwt_secret"] = "Secret key (leave blank to disable)";
 $strings["openByOnlyoffice"] = "Open with ONLYOFFICE";
 $strings["createNew"] = "Create new";
 $strings["title"] = "Title";