+++ lang/greek.php
@@ -1,12 +1,12 @@
 <?php
 /**
  * @author ASENSIO SYSTEM SIA
- * 
+ *
  */
 $strings["plugin_title"] = "ONLYOFFICE";
 $strings['plugin_comment'] = "Ο συνδετήρας ONLYOFFICE σάς επιτρέπει να προβάλετε, να επεξεργαστείτε και να συνεργαστείτε σε έγγραφα κειμένου, υπολογιστικά φύλλα και παρουσιάσεις στο Chamilo χρησιμοποιώντας έγγραφα ONLYOFFICE.";
-$strings["enableOnlyofficePlugin"] = "Ενεργοποίηση";
-$strings["documentServerUrl"] = "Διεύθυνση υπηρεσίας επεξεργασίας εγγράφων";
+$strings["enable_onlyoffice_plugin"] = "Ενεργοποίηση";
+$strings["document_server_url"] = "Διεύθυνση ONLYOFFICE Docs";
 $strings["openByOnlyoffice"] = "Άνοιγμα με ONLYOFFICE";
 $strings["createNew"] = "Δημιουργία νέου";
 $strings["title"] = "Τίτλος";