+++ lang/portuguese.php
@@ -1,12 +1,13 @@
 <?php
 /**
  * @author ASENSIO SYSTEM SIA
- * 
+ *
  */
 $strings["plugin_title"] = "ONLYOFFICE";
 $strings['plugin_comment'] = "ONLYOFFICE connector permite que você visualize, edite e colabore em documentos de texto, planilhas e apresentações dentro do Chamilo usando o ONLYOFFICE Docs.";
-$strings["enableOnlyofficePlugin"] = "Habilitar";
-$strings["documentServerUrl"] = "Endereço do serviço de edição de documentos";
+$strings["enable_onlyoffice_plugin"] = "Habilitar";
+$strings["document_server_url"] = "Endereço do ONLYOFFICE Docs";
+$strings["jwt_secret"] = "Chave secreta (deixe em branco para desativar)";
 $strings["openByOnlyoffice"] = "Abrir com ONLYOFFICE";
 $strings["createNew"] = "Criar novo";
 $strings["title"] = "Título";