+++ lib/langManager.php
@@ -17,16 +17,15 @@
  *
  */
 
-require_once __DIR__.'/../../../main/inc/global.inc.php';
+require_once __DIR__ . "/../../../main/inc/global.inc.php";
 
 class LangManager {
 
     /**
      * Return lang info for current user
-     * 
-     * @return array
      */
-    public static function getLangUser() {
+    public static function getLangUser(): array
+    {
         $langInfo = [];
         $userLang = api_get_language_from_type("user_profil_lang");
         if (empty($userLang)) {