+++ lib/onlyofficeActionObserver.php
@@ -22,10 +22,11 @@ class OnlyofficeActionObserver extends HookObserver implements HookDocumentActio
     /**
      * Constructor.
      */
-    public function __construct() {
+    public function __construct()
+    {
         parent::__construct(
-            'plugin/onlyoffice/onlyoffice_plugin.php',
-            'onlyoffice'
+            "plugin/onlyoffice/lib/onlyofficePlugin.php",
+            "onlyoffice"
         );
     }
 
@@ -34,11 +35,12 @@ class OnlyofficeActionObserver extends HookObserver implements HookDocumentActio
      *
      * @param HookDocumentActionEventInterface $event - the hook event
      */
-    public function notifyDocumentAction(HookDocumentActionEventInterface $event) {
+    public function notifyDocumentAction(HookDocumentActionEventInterface $event)
+    {
         $data = $event->getEventData();
 
-        if ($data['type'] === HOOK_EVENT_TYPE_PRE) {
-            $data['actions'][] = OnlyofficeTools::getButtonCreateNew();
+        if ($data["type"] === HOOK_EVENT_TYPE_PRE) {
+            $data["actions"][] = OnlyofficeTools::getButtonCreateNew();
             return $data;
         }
     }