+++ lib/onlyofficeTools.php
@@ -21,26 +21,23 @@ class OnlyofficeTools {
 
     /**
      * Return button-link to onlyoffice editor for file
-     * 
-     * @param array $document_data - document info
-     * 
-     * @return Display
      */
-    public static function getButtonEdit ($document_data) {
+    public static function getButtonEdit(array $document_data): string
+    {
 
         $plugin = OnlyofficePlugin::create();
 
-        $isEnable = $plugin->get("enableOnlyofficePlugin") === 'true';
+        $isEnable = $plugin->get("enable_onlyoffice_plugin") === "true";
         if (!$isEnable) {
-            return;
+            return '';
         }
 
         $urlToEdit = api_get_path(WEB_PLUGIN_PATH) . "onlyoffice/editor.php";
 
         $extension = strtolower(pathinfo($document_data["title"], PATHINFO_EXTENSION));
 
-        $canEdit = in_array($extension, FileUtility::$can_edit_types) ? true : false;
-        $canView = in_array($extension, FileUtility::$can_view_types) ? true : false;
+        $canEdit = in_array($extension, FileUtility::$can_edit_types);
+        $canView = in_array($extension, FileUtility::$can_view_types);
 
         $groupId = api_get_group_id();
         if (!empty($groupId)) {
@@ -53,22 +50,80 @@ class OnlyofficeTools {
         $urlToEdit = $urlToEdit . "docId=" . $documentId;
 
         if ($canEdit || $canView) {
-            return Display::url(Display::return_icon('../../plugin/onlyoffice/resources/onlyoffice_edit.png', $plugin->get_lang('openByOnlyoffice')), $urlToEdit);
+            return Display::url(
+                Display::return_icon(
+                    '../../plugin/onlyoffice/resources/onlyoffice_edit.png',
+                    $plugin->get_lang('openByOnlyoffice')
+                ),
+                $urlToEdit
+            );
         }
+
+        return '';
+    }
+
+    /**
+     * Return button-link to onlyoffice editor for view file
+     */
+    public static function getButtonView (array $document_data): string
+    {
+
+        $plugin = OnlyofficePlugin::create();
+
+        $isEnable = $plugin->get("enable_onlyoffice_plugin") === "true";
+        if (!$isEnable) {
+            return '';
+        }
+
+        $urlToEdit = api_get_path(WEB_PLUGIN_PATH) . "onlyoffice/editor.php";
+
+        $sessionId = api_get_session_id();
+        $courseInfo = api_get_course_info();
+        $documentId = $document_data["id"];
+        $userId = api_get_user_id();
+
+        $docInfo = DocumentManager::get_document_data_by_id($documentId, $courseInfo["code"], false, $sessionId);
+
+        $extension = strtolower(pathinfo($document_data["title"], PATHINFO_EXTENSION));
+        $canView = in_array($extension, FileUtility::$can_view_types);
+
+        $isGroupAccess = false;
+        $groupId = api_get_group_id();
+        if (!empty($groupId)) {
+            $groupProperties = GroupManager::get_group_properties($groupId);
+            $docInfoGroup = api_get_item_property_info(api_get_course_int_id(), 'document', $documentId, $sessionId);
+            $isGroupAccess = GroupManager::allowUploadEditDocument($userId, $courseInfo["code"], $groupProperties, $docInfoGroup);
+
+            $urlToEdit = $urlToEdit . "?groupId=" . $groupId . "&";
+        } else {
+            $urlToEdit = $urlToEdit . "?";
+        }
+
+        $isAllowToEdit = api_is_allowed_to_edit(true, true);
+        $isMyDir = DocumentManager::is_my_shared_folder($userId, $docInfo["absolute_parent_path"], $sessionId);
+
+        $accessRights = $isAllowToEdit || $isMyDir || $isGroupAccess;
+
+        $urlToEdit = $urlToEdit . "docId=" . $documentId;
+
+        if ($canView && !$accessRights) {
+            return Display::url(Display::return_icon('../../plugin/onlyoffice/resources/onlyoffice_view.png', $plugin->get_lang('openByOnlyoffice')), $urlToEdit, ["style" => "float:right; margin-right:8px"]);
+        }
+
+        return '';
     }
 
     /**
      * Return button-link to onlyoffice create new
-     * 
-     * @return Display
      */
-    public static function getButtonCreateNew () {
+    public static function getButtonCreateNew (): string
+    {
 
         $plugin = OnlyofficePlugin::create();
 
-        $isEnable = $plugin->get("enableOnlyofficePlugin") === 'true';
+        $isEnable = $plugin->get("enable_onlyoffice_plugin") === "true";
         if (!$isEnable) {
-            return;
+            return '';
         }
 
         $courseId = api_get_course_int_id();
@@ -76,12 +131,19 @@ class OnlyofficeTools {
         $groupId = api_get_group_id();
         $userId = api_get_user_id();
 
-        $urlToCreate = api_get_path(WEB_PLUGIN_PATH) . "onlyoffice/create.php?folderId=" . $_GET["id"] 
-                                                        . "&courseId=" . $courseId 
-                                                        . "&groupId=" . $groupId 
+        $urlToCreate = api_get_path(WEB_PLUGIN_PATH) . "onlyoffice/create.php"
+                                                        ."?folderId=" . (empty($_GET["id"])?'0':(int)$_GET["id"])
+                                                        . "&courseId=" . $courseId
+                                                        . "&groupId=" . $groupId
                                                         . "&sessionId=" . $sessionId
                                                         . "&userId=" . $userId;
 
-        return Display::url(Display::return_icon('../../plugin/onlyoffice/resources/onlyoffice_create.png', $plugin->get_lang('createNew')), $urlToCreate);
+        return Display::url(
+            Display::return_icon(
+                "../../plugin/onlyoffice/resources/onlyoffice_create.png",
+                $plugin->get_lang("createNew")
+            ),
+            $urlToCreate
+        );
     }
 }