+++ lib/templateManager.php
@@ -17,7 +17,7 @@
  *
  */
 
-require_once __DIR__.'/../../../main/inc/global.inc.php';
+require_once __DIR__ . "/../../../main/inc/global.inc.php";
 
 class TemplateManager {
 
@@ -52,12 +52,9 @@ class TemplateManager {
 
     /**
      * Return path to template new file
-     * 
-     * @param string $extension - extension of file
-     * 
-     * @return string
      */
-    public static function getEmptyTemplate($extension) {
+    public static function getEmptyTemplate($fileExtension): string
+    {
         $langInfo = LangManager::getLangUser();
         $lang = $langInfo["isocode"];
         if (!array_key_exists($lang, self::$localPath)) {
@@ -65,6 +62,6 @@ class TemplateManager {
         }
         $templateFolder = api_get_path(SYS_PLUGIN_PATH) . "onlyoffice/assets/" . self::$localPath[$lang];
 
-        return $templateFolder . "/" . ltrim($extension, ".") . ".zip";
+        return $templateFolder . "/" . ltrim($fileExtension, ".") . ".zip";
     }
 }