+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Change Log
 
+This plugin is developed and maintained at https://github.com/ONLYOFFICE/onlyoffice-chamilo.
+
+## 1.1.2
+- Add security filtering
+- Minor documentation and code style changes
+- View option for DOCX, XLSX, PPTX.
+- JWT support
+
 ## 1.0.0
-## Added
 - Ability to create documents
 - Edit option for DOCX, XLSX, PPTX.
 - Collaboration editing